@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\SampleConfig\SampleConfig.csproj" />
@@ -12,7 +12,6 @@
     <ProjectReference Include="..\..\src\Meceqs.AzureEventHubs\Meceqs.AzureEventHubs.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="idunno.Authentication.Basic" Version="2.1.1" />
   </ItemGroup>
 </Project>
