@@ -1,9 +1,9 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Meceqs.Serialization.Json
+namespace Meceqs.Serialization.NewtonsoftJson
 {
-    public static class JsonDefaults
+    public static class NewtonsoftJsonDefaults
     {
         public static readonly JsonSerializerSettings DefaultSerializerSettings = new JsonSerializerSettings
         {
